move ss.slurm -> ss.wlm.slurm, update tests & docs
